// client-sdk/models/ReportCommentsEntity.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CreateByEntity } from './CreateByEntity';
export type ReportCommentsEntity = {
    createdBy: CreateByEntity;
    id: number;
    createdById: number;
    createdAt: string;
    updatedAt: string;
    content: string;
    reportId: number;
};